+++ tests/logtalk/directives/use_module_2/plain.pl
@@ -2,7 +2,6 @@
 %
 %  This file is part of Logtalk <https://logtalk.org/>
 %  Copyright 1998-2021 Paulo Moura <pmoura@logtalk.org>
-%  SPDX-License-Identifier: Apache-2.0
 %
 %  Licensed under the Apache License, Version 2.0 (the "License");
 %  you may not use this file except in compliance with the License.
@@ -19,31 +18,6 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 
-:- if((
-	current_logtalk_flag(prolog_dialect, Dialect),
-	(Dialect == eclipse; Dialect == sicstus; Dialect == swi; Dialect == trealla; Dialect == yap)
-)).
-
-	:- initialization((
-		logtalk_load_context(directory, Directory),
-		atom_concat(Directory, module, Path),
-		use_module(Path)
-	)).
-
-	:- initialization((
-		set_logtalk_flag(report, warnings),
-		logtalk_load(plain),
-		logtalk_load(lgtunit(loader)),
-		logtalk_load(tests, [hook(lgtunit)]),
-		% the test/1 object parameter is used for testing
-		% predicate shortcuts declared in uses/2 directives
-		tests(2)::run
-	)).
-
-:- else.
-
-	:- initialization((
-		write('(not applicable)'), nl
-	)).
-
-:- endif.
+:- meta_predicate(bar(0)).
+bar(X) :-
+	call(X).